% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{CLLOmics}
\alias{CLLOmics}
\title{CLL multi-omic data}
\format{A list containing omic measurements for 92 patients
\itemize{
  \item{expr_lincRNA}{expression of lincRNA, p=1566}
  \item{expr_mRNA}{expression of mRNA, p=10910}
  \item{expr_mRNA}{expression of miRNA, p=1096}
  \item{met_3utr}{methylation in 3utr region, p=8661}
  \item{met_5utr}{methylation in 5utr region, p=8661}
  \item{met_cds_genebody}{methylation in coding region, p=9265}
  \item{met_intergenic}{methylation in intergenic region, p=11785}
  \item{met_noncds_genebody}{methylation in non-coding region, p=6884}
  \item{met_prom2k}{methylation in promotor region, p=12007}
  \item{surv1}{drug screen viability data, highest concentration, p=64}
  \item{surv2}{drug screen viability data, second-highest concentration, p=64}
  \item{surv3}{drug screen viability data, intermediate concentration, p=64}
  \item{surv4}{drug screen viability data, second-lowest concentration, p=64}
  \item{surv5}{drug screen viability data, lowest concentration, p=64}
}}
\usage{
data(CLLOmics)
}
\description{
A list containing differen omic measurements in CLL patient samples
}
