% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.grpRR.R
\name{coef.grpRR}
\alias{coef.grpRR}
\title{Get estimated coefficients from a grpRR object}
\usage{
\method{coef}{grpRR}(object, include_intercept = TRUE, ...)
}
\arguments{
\item{object}{fitted grpRR model as obtained from  \code{\link{grpRR}}}

\item{include_intercept}{whether to include the estimated intercept value in the output}

\item{...}{other arguments}
}
\value{
1-Column matrix of estimated coefficients.
}
\description{
Function to obtain estimated coefficients from a fitted grpRR model.
}
\examples{
# create data
dat <- makeExampleData()
fit <- grpRR(dat$X, dat$y, dat$annot)
coef(fit)
}
