% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{cv_compare}
\alias{cv_compare}
\title{Compare various regression method via cross-validation}
\usage{
cv_compare(X, y, annot, family = "gaussian", ncores = 1, nfolds = 10,
  plot_cv = FALSE, seed = NULL, parallel = FALSE, saveFits = FALSE,
  ...)
}
\arguments{
\item{X}{design matrix with samples in rows and features in columns (n x p)}

\item{y}{response vector of length n}

\item{annot}{factor of length p indicating group membership of each feature}

\item{family}{likelihood model for the response,
 either "gaussian" for linear regression
or "binomial" for logisitc regression}

\item{ncores}{number of cores to use}

\item{nfolds}{number of folds for evaluation}

\item{plot_cv}{whether to produce summary plots from evaluation}

\item{seed}{optional seed for the choice of folds}

\item{parallel}{whether to run cross-validation in parallel}

\item{saveFits}{whether to save the fit of each fold}

\item{...}{other parameters that can be passed to \code{\link{runMethods}}}
}
\value{
List with one entry per fold containing a list with
\describe{
\item{FPR}{False positive rate (requires \code{beta0} to be specified)}
\item{FNR}{False neagtive rate (requires \code{beta0} to be specified)}
\item{RMSE}{Root mean squared error on the left-out fold}
\item{pf_mat}{matrix with learnt penalty factors per group (rows) and method (column)}
\item{beta_mat}{matrix with estimated coefficients per feature (rows) and method (column)}
\item{intercepts}{estimated intercepts per method}
\item{sparsity_mat}{matrix with learnt sparsity levels per group (rows) and method (column) (0=sparse, 1=dense)}
\item{annot}{annotation of features to groups as specified when calling \code{\link{cv_compare}}}
\item{runtime}{vector of runtimes for the different methods}
\item{l1error_intercept}{absolute error on the estimated intercept  (requires \code{trueintercept} to be specified)}
\item{l1error_beta}{absolute error on the estimated coefficients (requires \code{beta0} to be specified)}
}
}
\description{
Function to fit a regresion model using serveral different methods
and evaluate them in a cross-validated fashion in terms of prediction and estimation performance
}
\details{
This function can be used to test various method for regression on a dataset in a cross-validated fashion.
It fits the methods on all except one fold and evaluates their prediction performance on the remaining fold.
If the true coefficients of the model are known they can be specified via \code{beta0} and \code{trueintercept}.
Then, additionally the error on the estimates is evaluates as well as the feature selection performance.
Note that for grpRR the selected features are determined by the posterior inclusion probabilities  with
a feature being called active for s>0.5 (The method grpRR_cutoff sets inactive with s<=0.5 coefficients to exactly zero).
}
\examples{
dat <- makeExampleData()
cv.out <- cv_compare(dat$X, dat$y, dat$annot, nfolds=3)
}
