% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{evalResult}
\alias{evalResult}
\title{evalResult}
\usage{
evalResult(allFits, Xtest, ytest, plotit = T, family = "gaussian",
  plotbeta = F, saveit = F, filenm = "")
}
\arguments{
\item{allFits}{List as produced by \code{\link{runMethods}}}

\item{Xtest}{Design matrix of size n' x p (same feature structure as used in runMethods)}

\item{ytest}{Response vector of size n'}

\item{plotit}{Boolean, indication wether to produce plots of MSE and penalty factors}
}
\value{
Datframe containing error measures per method
}
\description{
Function to evaluate results (MSE, Peanlty Factors...)
}
