% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_by_MCMC.R
\name{fit_by_MCMC}
\alias{fit_by_MCMC}
\title{fit_by_MCMC}
\usage{
fit_by_MCMC(X, y, annot, iter = 12000, warmup = 2000, chains = 3,
  verbose = T)
}
\arguments{
\item{X}{Design matrix of size n x p}

\item{y}{Response vector of size n}

\item{annot}{Factor of length p indicating group membership of each feature}

\item{iter}{maximum number of iterations}

\item{warmup}{number of iterations for burn-in period}

\item{chains}{number of chains to run}

\item{...}{other parameters that can be passed to stan}
}
\value{
SummaryList containing model coefficients, penalty factors, run time, fit model
}
\description{
Function to fit the hier. Bayes model used for grpRR using MCMC for inference. Only implemented for the normal prior and gaussian response at the moment.
}
