% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_grpRR.R
\name{fit_grpRR}
\alias{fit_grpRR}
\title{Fit grpRR model}
\usage{
fit_grpRR(X, y, annot, factoriseQ = TRUE, spikeslab = TRUE, d_tau = 0.001,
  r_tau = 0.001, d_gamma = 0.001, r_gamma = 0.001, r_pi = 1, d_pi = 1,
  max_iter = 1000, th = 1e-05, intercept = TRUE, calcELB = TRUE,
  verbose = TRUE, freqELB = 10, family = "gaussian", nogamma = F,
  standardize = TRUE)
}
\arguments{
\item{X}{Design matrix of size n x p}

\item{y}{Response vector of size n}

\item{annot}{Factor of length p indicating group membership of each feature}

\item{factoriseQ}{If true, the variational distribution is assumed to fully factorize across features (rougher approx., but faster). If spikeslab=F, this is always done.}

\item{spikeslab}{If true, a spike and slab prior is used instead of a normal prior}

\item{d_tau}{hyper-parameters for prior of tau (noise precision)}

\item{r_tau}{hyper-parameters for prior of tau (noise precision)}

\item{d_gamma}{hyper-parameters for prior of gamma (coeffient's prior precision)}

\item{r_gamma}{hyper-parameters for prior of gamma (coeffient's prior precision)}

\item{r_pi}{hyper-parameters for prior of pi (Spike-Slab-Bernoulli variable prior probabiliy of being 1)}

\item{d_pi}{hyper-parameters for prior of pi (Spike-Slab-Bernoulli variable prior probabiliy of being 1)}

\item{max_iter}{maximum number of iterations}

\item{th}{convergence threshold for ELB}

\item{intercept}{boolean, indicating wether to fit an intercept}

\item{calcELB}{boolean, indicating wether to calculate ELB}

\item{verbose}{boolean, indicating wether to print out intermediate messages during fitting}

\item{freqELB}{determines frequency at which ELB is to be calculated, i.e. each feqELB-th iteration}

\item{nogamma}{If true, the normal prior will have same variance for all groups (only relevant for SS)}
}
\value{
List of fitted parameters .....
}
\description{
Main function to fit a grpRR model.
}
