% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{gdsc1000}
\alias{gdsc1000}
\title{GSDC multi-omic data}
\format{Including for n=962 samples
\itemize{
  \item{auc}{Aera under the curve for p=265 drugs}
  \item{BEM}{Binary Event Matrix, p=1250 (hypermethylation, mutations, CNA)}
  \item{CellLinesCommon}{Info on cell lines}
  \item{DrugMeta}{Info on drugs}
  \item{expr}{RMA normalized expression, p=17737}
  \item{exprgenesMeta}{Info on Genes in Expression data}
  \item{ic50}{IC50 values for p=265 drugs}
  \item{speedScoreBinary}{Speed Score from expression data, p=22}
  \item{X}{joint design matrix, p=19009}
  \item{annot}{annotation to 5 groups 'mut'   'meth'  'CNA'   'expr'  'speed' }
}}
\usage{
data(gdsc1000)
}
\description{
A Rdata object containing differen omic measurements in cancer cell lines (see Iorio et al)
}
