% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromSummary.R
\name{getBS}
\alias{getBS}
\title{get Brier Score}
\usage{
getBS(AllFits)
}
\arguments{
\item{AllFits}{object as produced by \code{\link{evaluateFits}}}
}
\value{
A vector of the Brier score on the test data for each method.
}
\description{
Function to obtain the Brier Score of various methods in a logistic regression model.
}
\examples{
dat <- makeExampleData(response = "bernoulli")
ntrain <- dat$n/2
fit <- runMethods(dat$X[seq_len(ntrain),], dat$y[seq_len(ntrain)], dat$annot, family = "binomial")
evalFit <- evaluateFits(fit, dat$X[seq_len(ntrain) + dat$n/2,],
 dat$y[seq_len(ntrain)+ dat$n/2])
getBS(evalFit)
}
