% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromSummary.R
\name{getCoefficients}
\alias{getCoefficients}
\title{get coefficients}
\usage{
getCoefficients(AllFits)
}
\arguments{
\item{AllFits}{object as produced by \code{\link{runMethods}}}
}
\value{
A matrix of coefficients per feature (rows) and method (columns).
}
\description{
get coefficients estimated from various methods
}
\examples{
dat <- makeExampleData()
fit <- runMethods(dat$X, dat$y, dat$annot)
getCoefficients(fit)
}
