% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromSummary.R
\name{getFNR}
\alias{getFNR}
\title{get FNR}
\usage{
getFNR(AllFits)
}
\arguments{
\item{AllFits}{object as produced by \code{\link{evaluateFits}}}
}
\description{
get false negative rates of various methods
}
