% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromSummary.R
\name{getFPR}
\alias{getFPR}
\title{get FPR}
\usage{
getFPR(AllFits)
}
\arguments{
\item{AllFits}{object as produced by \code{\link{evaluateFits}}}
}
\value{
A vector of the FPR on the selected coefficients for each method.
}
\description{
get false positive rates of various methods
}
\examples{
dat <- makeExampleData()
ntrain <- dat$n/2
fit <- runMethods(dat$X[seq_len(ntrain),],
 dat$y[seq_len(ntrain)], dat$annot, beta0 = dat$beta)
evalFit <- evaluateFits(fit, dat$X[seq_len(ntrain) + dat$n/2,],
 dat$y[seq_len(ntrain)+ dat$n/2])
getFPR(evalFit)
}
