% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromSummary.R
\name{getl1error_beta}
\alias{getl1error_beta}
\title{get l1 error on beta}
\usage{
getl1error_beta(AllFits)
}
\arguments{
\item{AllFits}{object as produced by \code{\link{evaluateFits}}}
}
\value{
A vector of the absolute error on the estimated coefficients for each method.
}
\description{
get absolute error on beta
}
\examples{
dat <- makeExampleData()
ntrain <- dat$n/2
fit <- runMethods(dat$X[seq_len(ntrain),],
 dat$y[seq_len(ntrain)], dat$annot, beta0 = dat$beta)
evalFit <- evaluateFits(fit, dat$X[seq_len(ntrain) + dat$n/2,],
 dat$y[seq_len(ntrain)+ dat$n/2])
getl1error_beta(evalFit)
}
