% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromSummary.R
\name{getl1error_intercept}
\alias{getl1error_intercept}
\title{get l1 error on intercept}
\usage{
getl1error_intercept(AllFits)
}
\arguments{
\item{AllFits}{object as produced by \code{\link{evaluateFits}}}
}
\description{
get absolute error on intercept
}
