% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMarginalCoefficient}
\alias{getMarginalCoefficient}
\title{Get marginal coefficient from a (generalized) linear model}
\usage{
getMarginalCoefficient(response, data, family = "gaussian")
}
\arguments{
\item{response}{response vector}

\item{data}{design matrix}

\item{family}{liklihood model for the response}
}
\value{
a matrix containing estimates, standard errors, statistics
 and p-values (rows) for each feature (columns) in the desing matrix
}
\description{
Function to compute marginal regression coefficients using \code{glm}.
}
\examples{
dat <- makeExampleData(response="bernoulli")
getMarginalCoefficient(dat$y, dat$X, family= "binomial")
}
