% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromSummary.R
\name{getMSE}
\alias{getMSE}
\title{get MSE}
\usage{
getMSE(AllFits)
}
\arguments{
\item{AllFits}{object as produced by \code{\link{evaluateFits}}}
}
\value{
A vector of the mean squared error on the test data for each method.
}
\description{
get MSE of various methods
}
\examples{
dat <- makeExampleData()
ntrain <- dat$n/2
fit <- runMethods(dat$X[seq_len(ntrain),],
 dat$y[seq_len(ntrain)], dat$annot)
evalFit <- evaluateFits(fit, dat$X[seq_len(ntrain) + dat$n/2,],
 dat$y[seq_len(ntrain)+ dat$n/2])
getMSE(evalFit)
}
