% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getPIPs}
\alias{getPIPs}
\title{Get posterior inclusion probabilities per feature}
\usage{
getPIPs(object)
}
\arguments{
\item{object}{fitted grpRR model as obtained from  \code{\link{grpRR}}}
}
\value{
1-Column matrix of estimated posterior inclusion probabilities.
}
\description{
Function to obtain estimated posterior inclusion probabilities per feature from a fitted grpRR model.
}
\examples{
# create data
dat <- makeExampleData()
fit <- grpRR(dat$X, dat$y, dat$annot)
getPIPs(fit)
}
