% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromSummary.R
\name{getRMSE}
\alias{getRMSE}
\title{get RMSE}
\usage{
getRMSE(AllFits)
}
\arguments{
\item{AllFits}{object as produced by \code{\link{evaluateFits}}}
}
\value{
A vector of the root mean squared error on the test data for each method.
}
\description{
get RMSE of various methods
}
\examples{
dat <- makeExampleData()
ntrain <- dat$n/2
fit <- runMethods(dat$X[seq_len(ntrain),],
 dat$y[seq_len(ntrain)], dat$annot)
evalFit <- evaluateFits(fit, dat$X[seq_len(ntrain) + dat$n/2,],
 dat$y[seq_len(ntrain)+ dat$n/2])
getRMSE(evalFit)
}
