% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromSummary.R
\name{getRunTime}
\alias{getRunTime}
\title{get run times}
\usage{
getRunTime(AllFits)
}
\arguments{
\item{AllFits}{object as produced by \code{\link{runMethods}}}
}
\value{
A vector of run times for each method.
}
\description{
Function to obtain the runtimes of various methods from the fitted regression models.
}
\examples{
dat <- makeExampleData()
fit <- runMethods(dat$X, dat$y, dat$annot)
getRunTime(fit)
}
