% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{makeExampleData}
\alias{makeExampleData}
\title{Simulate example data with groups of equal size}
\usage{
makeExampleData(n = 100, p = 200, g = 4, gammas = c(0.1, 1, 10,
  100), pis = c(0.5, 0.5, 0.5, 0.5), tau = 1, rho = 0,
  response = "gaussian", intercept = 0)
}
\arguments{
\item{n}{number of samples}

\item{p}{number of features}

\item{g}{number of groups}

\item{gammas}{vector of length g, sepcifying the slab precision of the prior on beta per group}

\item{pis}{vector of length g, sepcifying the probability of s to be 1 (slab)}

\item{tau}{noise precision}

\item{rho}{correlation of design matrix (Toeplitz structure)}

\item{response}{gaussion or bernoulli for linear or logistic regression, respectively}

\item{intercept}{model intercept (default: 0)}
}
\value{
list containin the design matrix X, the response y, the feautre annotation to
 groups annot as well as the different parameters in the Bayesian model and the correlation strength rho
}
\description{
Simulate data from the Bayesian model with specified parameters gamma and tau
}
