% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MarginalCoefficient}
\alias{MarginalCoefficient}
\title{MarginalCoefficient}
\usage{
MarginalCoefficient(response, data, family = "gaussian")
}
\description{
Function to compute marignal regression coefficients
}
