% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotELBO}
\alias{plotELBO}
\title{Plot evidence lower bound}
\usage{
plotELBO(fit)
}
\arguments{
\item{fit}{fit as produced by \code{\link{grpRR}}}
}
\value{
a ggplot object
}
\description{
Function to plot the evidence lower bound (ELBO) over iterations to monitor the convergence of the algorithm.
}
\examples{
dat <- makeExampleData()
fit <- grpRR(dat$X, dat$y, dat$annot)
plotELBO(fit)
}
