% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotMethodComparison}
\alias{plotMethodComparison}
\title{Plot performance measures for various methods}
\usage{
plotMethodComparison(resultList, family = "gaussian",
  methods2plot = "all")
}
\arguments{
\item{resultList}{list as created by \code{\link{compareMethodsCV}}}

\item{family}{gaussian or binomial (same as used in \code{\link{compareMethodsCV}})}

\item{methods2plot}{which methods to include into the fit}
}
\value{
a ggplot object
}
\description{
Function to plot performance measures of the fits obtained from various regression methods that were evaluated using to \code{\link{compareMethodsCV}}.
}
\examples{
dat <- makeExampleData()
cv.out <- compareMethodsCV(dat$X, dat$y, dat$annot, nfolds=3)
plotMethodComparison(cv.out)
}
