% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.grpRR.R
\name{print.grpRR}
\alias{print.grpRR}
\title{Print a grpRR object}
\usage{
\method{print}{grpRR}(x, ...)
}
\arguments{
\item{x}{fitted grpRR model as obtained from  \code{\link{grpRR}}}

\item{...}{additional print arguments}
}
\value{
Print output.
}
\description{
Function to print a fitted grpRR model.
}
\examples{
# create data
dat <- makeExampleData()
fit <- grpRR(dat$X, dat$y, dat$annot)
print(fit)
}
