% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{runMethods}
\alias{runMethods}
\title{Run various regression methods}
\usage{
runMethods(Xtrain, ytrain, annot, beta0 = NULL, trueintercept = NULL,
  max_iter = 5000, family = "gaussian", intercept = TRUE,
  standardize = TRUE, freqELB = 10, calcELB = TRUE, th = 0.01,
  n_rep = 1, verbose = FALSE, verbose_progress = TRUE,
  compareGRridge = FALSE, include_nonfacQ = FALSE,
  compareSparseGroupLasso = FALSE, compareIPF = FALSE,
  compareGroupLasso = FALSE, includeRF = FALSE,
  compareAdaLasso = FALSE, include_varbvs = FALSE,
  include_nogamma = FALSE, include_grpRR_SS_ungrouped = FALSE)
}
\arguments{
\item{Xtrain}{Design matrix of size n x p}

\item{ytrain}{Response vector of size n}

\item{annot}{Factor of length p indicating group membership of each feature}

\item{beta0}{True coefficients in the linear model if known,
NULL otherwise (default)}

\item{trueintercept}{True intercept in the linear model if known,
NULL otherwise (default)}

\item{max_iter}{Maximum number of iterations for grpRR methods (see also  \code{\link{fit_grpRR}})}

\item{family}{Likelihood model for the response,
 either "gaussian" for linear regression
or "binomial" for logisitc regression}

\item{intercept}{boolean, indicating wether to fit an intercept}

\item{standardize}{boolean, indicating wether
features for Ridge and Lasso fit should be standardized.
Note this does not affect GRridge and group Lasso
 where standardization is default.}

\item{freqELB}{Frequency at which the evidence lower bound (ELB) is to be calculated for grpRR,
i.e. each freqELB-th iteration (see also  \code{\link{fit_grpRR}})}

\item{calcELB}{Boolean, indicating whether to calculate the evidence lower bound (ELB) for grpRR (see also  \code{\link{fit_grpRR}})}

\item{th}{Convergence threshold for the evidence lower bound (ELB) in grpRR (see also  \code{\link{fit_grpRR}})}

\item{n_rep}{Number of reptitions with different random initilizations to be fit in grpRR (see also  \code{\link{fit_grpRR}})}

\item{verbose}{Boolean, indicating wether
to print out intermediate messages during fitting}

\item{verbose_progress}{Boolean, indicating
whether to print details on the progress}

\item{compareGRridge}{Boolean, indicating
wether to fit a GRridge model}

\item{include_nonfacQ}{include a VB method
with multivariate variational distributon
(can be very timme consuming for large data sets)}

\item{compareSparseGroupLasso}{Boolean, indicating
whether to fit a sparse group lasso}

\item{compareIPF}{Boolean, indicating
whether to fit a IPFLasso}

\item{compareGroupLasso}{Boolean, indicating
whether to fit a group lasso}

\item{includeRF}{Boolean, indicating
whether to fit a random forest}

\item{compareAdaLasso}{Boolean, indicating
whether to fit an adpative lasso}

\item{include_varbvs}{Boolean, indicating
whether to fit varbvs}

\item{include_nogamma}{Boolean, indicating
whether to fit a grpRR model without different slab parameters}

\item{include_grpRR_SS_ungrouped}{Boolean, indicating
whether to fit a grpRR model without group annotations}
}
\value{
List of fitted models and two data frames with
 coeffcients and penalty factors

a list with
 - a summaryList containing the complete individual fits
   as well as other statistics such as estimated coefficients,
    runtime, sparsity, intercept, penalty factors)
- the details on the data (sample size n, predictor number p,
 covariate annotation annot, group number G,
  feature and annotation names)
}
\description{
Function to fit a regression methods using serveral different methods.
}
\examples{
dat <- makeExampleData()
allFits <- runMethods(Xtrain=dat$X,
 ytrain=dat$y, annot=dat$annot)
}
