% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_grpLM}
\alias{simulate_grpLM}
\title{Simulate Data According to a linear model with grouped feautres}
\usage{
simulate_grpLM(n = 100, p = 1000, beta_best = 3, G = 10,
  block_cor = 0.3, blockSize = 10, equiCor = 0.1, SLinG = 0.7,
  SLovG = 0.5, model = "linear", diffFac = 0.7, seed = 167,
  sigma2 = 0.1, family = "gaussian", onlypos = T)
}
\arguments{
\item{n}{number of samples}

\item{p}{number of features}

\item{beta_best}{coefficient amplitude in most-informative group}

\item{G}{number of groups of features}

\item{block_cor}{overall correlation between all features}

\item{blockSize}{block size of correlated blocks}

\item{equiCor}{coefficient amplitude in most-informative group}

\item{SLinG}{sparsity level within each group, i.e. percentage of zero feautres within group}

\item{SLovG}{(overall) sparsity level, i.e. percentage of zero groups}

\item{model}{model for response (e.g. linear, based on logs)}

\item{diffFac}{differential signal per group \eqn{\beta_{g+1} = \beta_g *diffFac}}

\item{sigma2}{noise variance}

\item{family}{family of response, e.g. gaussian, binomial}

\item{onlypos}{boolean determining wether within a group all non-zero coeffcients should have the same postive value (if TRUE) or wether half get a negative sign (if FALSE)}
}
\value{
A list containing
\item{X}{nxp matrix of feautres with given correlation structure}
\item{y}{vector of length p of simulated response}
\item{annot}{factor containing group assignemnt of featrues}
\item{beta0}{true coefficients in the model}
\item{family}{family used for response}
}
\description{
The predictors are split into group carrying different amount of information (different coeffcient sizes)
The simulation is based on multivariate normal distributed predictors.
}
\details{
Special cases would be orthogonal design (blockSize=1,equiCor=0) or equicorrelated design (blockSize=1, equiCor=rho).
}
