% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulateData_toeplitz}
\alias{simulateData_toeplitz}
\title{Simulate Data According to a linear model with given coefficients and toeplitz design matrix}
\usage{
simulateData_toeplitz(n, p, beta, sigma2, seed, rho)
}
\description{
Simulate Data According to a linear model with given coefficients and toeplitz design matrix
}
