% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulateExplicit}
\alias{simulateExplicit}
\title{Simulate Data According to a linear model with given coefficients}
\usage{
simulateExplicit(n, p, beta, sigma2, seed, block_cor = 0, equiCor = 0,
  blockSize = 0)
}
\description{
Simulate Data According to a linear model with given coefficients
}
