% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{cv_compare}
\alias{cv_compare}
\title{Compare various regression method via cross-validation}
\usage{
cv_compare(X, y, annot, family = "gaussian", ncores = 1, nfolds = 10,
  plot_cv = FALSE, seed = NULL, parallel = FALSE, saveFits = FALSE,
  ...)
}
\arguments{
\item{X}{Design matrix of size n x p}

\item{y}{Response vector of size n}

\item{annot}{Factor of length p indicating group membership of each feature}

\item{family}{Liklihood model to use for response, either binomial or gaussian}

\item{ncores}{Number of cores to use}

\item{nfolds}{Number of fold for evaluation}

\item{plot_cv}{boolean whether to plot summary from evaluation}

\item{seed}{optional seed for the choice of folds}

\item{parallel}{boolean: Run cross-validation in parallel?}

\item{saveFits}{boolean: Save the fit of each fold?}

\item{...}{Other parameters that can be passed to \code{\link{runMethods}}}
}
\value{
List of perfromnce measures in each fold
}
\description{
Function to run serveral different methods for
high-dimensional regression and evaluate them in a cross-validated fashion
}
\examples{
dat <- makeExampleData()
cv.out <- cv_compare(dat$X, dat$y, dat$annot, nfolds=3)
}
