% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{cv_compare}
\alias{cv_compare}
\title{Compare various regression method via cross-validation}
\usage{
cv_compare(X, y, annot, family = "gaussian", ncores = 1, nfolds = 10,
  plot_cv = FALSE, seed = NULL, parallel = FALSE, saveFits = FALSE,
  ...)
}
\arguments{
\item{X}{design matrix with samples in rows and features in columns (n x p)}

\item{y}{response vector of length n}

\item{annot}{factor of length p indicating group membership of each feature}

\item{family}{likelihood model for the response,
 either "gaussian" for linear regression
or "binomial" for logisitc regression}

\item{ncores}{number of cores to use}

\item{nfolds}{number of fold for evaluation}

\item{plot_cv}{whether to produce summary plots from evaluation}

\item{seed}{optional seed for the choice of folds}

\item{parallel}{whether to run cross-validation in parallel}

\item{saveFits}{whether to save the fit of each fold}

\item{...}{other parameters that can be passed to \code{\link{runMethods}}}
}
\value{
List of perfromnce measures in each fold
}
\description{
Function to fit a regresion model using serveral different methods
and evaluate them in a cross-validated fashion in terms of prediction and estimation performance
}
\examples{
dat <- makeExampleData()
cv.out <- cv_compare(dat$X, dat$y, dat$annot, nfolds=3)
}
