% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_compare.R
\name{cv_compare}
\alias{cv_compare}
\title{cv_compare}
\usage{
cv_compare(X, y, annot, family = "gaussian", ncores = 1, nfolds = 10,
  plot_cv = TRUE, seed = NULL, parallel = FALSE, saveFits = FALSE, ...)
}
\arguments{
\item{X}{Design matrix of size n x p}

\item{y}{Response vector of size n}

\item{annot}{Factor of length p indicating group membership of each feature}

\item{family}{Liklihood model to use for response, either binomial or gaussian}

\item{ncores}{Number of cores to use}

\item{nfolds}{Number of fold for evaluation}

\item{plot_cv}{boolean whether to plot summary from evaluation}

\item{seed}{optional seed for the choice of folds}

\item{...}{Other parameters that can be passed to RunMethods}
}
\value{
List of fitted models and two data frames with coeffcients and penalty factors
}
\description{
Function to run serveral different methods for high-dimensional regression and evaluate them in a cross-validated fashion
}
