% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{evaluateFits}
\alias{evaluateFits}
\title{Evaluate fits from various regression methods}
\usage{
evaluateFits(allFits, Xtest, ytest)
}
\arguments{
\item{allFits}{List as produced by \code{\link{runMethods}}}

\item{Xtest}{Design matrix of size n' x p (same feature structure as used in \code{\link{runMethods}})}

\item{ytest}{Response vector of size n'}
}
\value{
List as prodcused by \code{\link{runMethods}} with additional predicition performance slots
}
\description{
Function to evaluate results on test data
}
