% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{evaluateModel}
\alias{evaluateModel}
\title{Evaluate a fitted regression model on test data}
\usage{
evaluateModel(beta_est, intercept, X_test, y_test, beta0 = NULL, family)
}
\arguments{
\item{beta_est}{estimated model coefficients}

\item{intercept}{estimated intercept}

\item{X_test}{predictor matrix of test data of size
number of test samples (n_test) times features (p)}

\item{y_test}{response vector of length number of test samples (n_test)}

\item{beta0}{true model coefficients (if known)}

\item{family}{likelihood model for the response, either
"gaussian" for linear regression or "binomial" for logistic regression}
}
\value{
A list containing various performance measures such as RMSE, FNR, FPR etc.
}
\description{
Function to evaluate a fitted regression model on test data in terms of their prediction and feature selection properties.
}
\examples{
dat <- makeExampleData()
# take a null model with all coefficeints set to zero for evaluation
beta <- rep(0, dat$p)
eval.out <- evaluateModel(beta, intercept = 0, X_test = dat$X,
 y_test = dat$y, beta0 = dat$beta, family = "gaussian")
}
