% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromSummary.R
\name{getCoefficients}
\alias{getCoefficients}
\title{get coefficients}
\usage{
getCoefficients(AllFits)
}
\arguments{
\item{AllFits}{object as produced by RunMethods}
}
\description{
get coefficients estimated from various methods
}
