% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromSummary.R
\name{getFPR}
\alias{getFPR}
\title{get FPR}
\usage{
getFPR(AllFits)
}
\arguments{
\item{AllFits}{object as produced by \code{\link{evaluateFits}}}
}
\description{
get false positive rates of various methods
}
