% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromSummary.R
\name{getIntercept}
\alias{getIntercept}
\title{get intercept}
\usage{
getIntercept(AllFits)
}
\arguments{
\item{AllFits}{object as produced by \code{\link{runMethods}}}
}
\value{
A vector of intercept values for each method.
}
\description{
get intercept estimated from various methods
}
\examples{
dat <- makeExampleData()
fit <- runMethods(dat$X, dat$y, dat$annot)
getIntercept(fit)
}
