% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromSummary.R
\name{getIntercept}
\alias{getIntercept}
\title{get intercept}
\usage{
getIntercept(AllFits)
}
\arguments{
\item{AllFits}{object as produced by RunMethods}
}
\description{
get intercept estimated from various methods
}
