% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromSummary.R
\name{getl1error_beta}
\alias{getl1error_beta}
\title{get l1 error on beta}
\usage{
getl1error_beta(AllFits)
}
\arguments{
\item{AllFits}{object as produced by \code{\link{evaluateFits}}}
}
\description{
get absolute error on beta
}
