% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromSummary.R
\name{getPenaltyFactors}
\alias{getPenaltyFactors}
\title{get penalty factors}
\usage{
getPenaltyFactors(AllFits)
}
\arguments{
\item{AllFits}{object as produced by \code{\link{runMethods}}}
}
\value{
A matrix of penalty factors per group (rows) and method (columns).
}
\description{
get penalty factors from various methods
}
\examples{
dat <- makeExampleData()
fit <- runMethods(dat$X, dat$y, dat$annot)
getPenaltyFactors(fit)
}
