% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromSummary.R
\name{getPenaltyFactors}
\alias{getPenaltyFactors}
\title{get penalty factors}
\usage{
getPenaltyFactors(AllFits)
}
\arguments{
\item{AllFits}{object as produced by RunMethods}
}
\description{
get penalty factors from various methods
}
