% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromSummary.R
\name{getSparsityLevel}
\alias{getSparsityLevel}
\title{get sparsity levels}
\usage{
getSparsityLevel(AllFits)
}
\arguments{
\item{AllFits}{object as produced by RunMethods}
}
\description{
get sparsity levels (1=dense, 0=sparse) from various methods
}
