% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromSummary.R
\name{getSparsityLevel}
\alias{getSparsityLevel}
\title{get sparsity levels}
\usage{
getSparsityLevel(AllFits)
}
\arguments{
\item{AllFits}{object as produced by \code{\link{runMethods}}}
}
\value{
A matrix of sparsity levels per group (rows) and method (columns).
}
\description{
Function to obtain the sparsity levels (1=dense, 0=sparse) estimated by various methods from the fitted regression models.
}
\examples{
dat <- makeExampleData()
fit <- runMethods(dat$X, dat$y, dat$annot)
getSparsityLevel(fit)
}
