% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{imputeByMean}
\alias{imputeByMean}
\title{Impute by mean}
\usage{
imputeByMean(x)
}
\arguments{
\item{x}{vector to with missing values to impute}
}
\value{
vector with imputed values
}
\description{
Function to impute missing values of a vector by its mean.
}
\examples{
x <- c(1,0.4,1.3, NA)
imputeByMean(x)
}
