% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{makeExampleDataWithUnequalGroups}
\alias{makeExampleDataWithUnequalGroups}
\title{makeExampleDataWithUnequalGroups}
\usage{
makeExampleDataWithUnequalGroups(n = 100, pg = c(100, 100, 10, 10),
  gammas = c(0.1, 10, 0.1, 10), pis = c(0.5, 0.5, 0.5, 0.5), tau = 1,
  rho = 0)
}
\arguments{
\item{n}{number of samples}

\item{pg}{vector of length g (desired number of groups) with number of features per group}

\item{gammas}{vector of length g, sepcifying the slab precision of the prior on beta per group}

\item{pis}{vector of length g, sepcifying the probability of s to be 1 (slab)}

\item{tau}{noise precision}

\item{rho}{correlation of design matrix (Toeplitz structure)}
}
\description{
Simulate data from the Bayesian model with specified parameters $\gamma$ and $\tau$.
}
