% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotELBO}
\alias{plotELBO}
\title{plotELBO}
\usage{
plotELBO(fit)
}
\arguments{
\item{fit}{fit as produced by fit_grpRR}
}
\description{
plot the ELBO over iteration to monitor convergence of the algorithm
}
