% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotGroupPenalties}
\alias{plotGroupPenalties}
\title{Plot group-wise penalties}
\usage{
plotGroupPenalties(fit)
}
\arguments{
\item{fit}{fit as produced by \code{\link{grpRR}}}
}
\value{
a ggplot object
}
\description{
Function to plot the group-wise penalty factors (gamma) and sparsity levels.
}
\examples{
dat <- makeExampleData()
fit <- grpRR(dat$X, dat$y, dat$annot)
plotELBO(fit)
}
