% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotMethodComparison}
\alias{plotMethodComparison}
\title{@title plot comparison of methods
 @name plotMethodComparison
 @description Function to plot method comparison across several runs}
\usage{
plotMethodComparison(resultList, family = "gaussian",
  methods2plot = "all")
}
\arguments{
\item{resultList}{List as created by cv_compare}

\item{family}{gaussian or binomial (same as used in cv_compare)}

\item{methods2plot}{which method to be plotted}
}
\description{
@title plot comparison of methods
 @name plotMethodComparison
 @description Function to plot method comparison across several runs
}
