% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{plotMethodComparison}
\alias{plotMethodComparison}
\title{#'  plotMethodComparison
#'
Function to plot method comparison across several runs}
\usage{
plotMethodComparison(resultList)
}
\arguments{
\item{resultList}{List as in simulation_setting1.Rmd}
}
\description{
#'  plotMethodComparison
#'
Function to plot method comparison across several runs
}
