% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotPosterior}
\alias{plotPosterior}
\title{plotPosterior}
\usage{
plotPosterior(fit, param2plot, beta0 = NULL, gamma0 = NULL,
  tau0 = NULL, pi0 = NULL, s0 = NULL, jmax = 2, range = NULL)
}
\arguments{
\item{fit}{fit as produced by fit_grpRR}

\item{param2plot}{which parametet to plot (gamma, beta, tau or s)}

\item{beta0}{true beta (if known)}

\item{gamma0}{true gamma (if known)}

\item{tau0}{true tau (if known)}

\item{pi0}{true pi (if known)}

\item{s0}{true s (if known)}

\item{jmax}{maximal number of betas per group to plot}

\item{range}{plotting range (x-axis)}
}
\value{
a ggplot object
}
\description{
plot the posterior of the model parameters
}
\examples{
dat <- makeExampleData()
fit <- fit_grpRR(dat$X, dat$y, dat$annot)
plotPosterior(fit, param2plot="gamma")
}
