% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{runMethods}
\alias{runMethods}
\title{Run various regression methods}
\usage{
runMethods(Xtrain, ytrain, annot, family = "gaussian",
  intercept = TRUE, standardize = TRUE, beta0 = NULL,
  trueintercept = NULL, max_iter = 5000, freqELB = 10,
  calcELB = TRUE, th = 0.01, n_rep = 1, verbose = FALSE,
  verbose_progress = TRUE, compareGRridge = FALSE,
  include_nonfacQ = FALSE, compareSparseGroupLasso = FALSE,
  compareIPF = FALSE, compareGroupLasso = FALSE, includeRF = FALSE,
  compareAdaLasso = FALSE, include_varbvs = FALSE,
  include_nogamma = FALSE, include_grpRR_SS_ungrouped = FALSE)
}
\arguments{
\item{Xtrain}{design matrix with samples in rows and features in columns (n x p)}

\item{ytrain}{response vector of length n}

\item{annot}{factor of length p indicating group membership of each feature}

\item{family}{likelihood model for the response,
 either "gaussian" for linear regression
or "binomial" for logisitc regression}

\item{intercept}{whether to include an intercept into the model}

\item{standardize}{whether to standardize the predictors to unit variance.
Note this does not affect GRridge and group Lasso  where standardization is default.}

\item{beta0}{true coefficients in the linear model if known,
NULL otherwise (default)}

\item{trueintercept}{true intercept in the linear model if known,
NULL otherwise (default)}

\item{max_iter}{maximum number of iterations for grpRR methods (see also  \code{\link{fit_grpRR}})}

\item{freqELB}{frequency at which the evidence lower bound (ELB) is to be calculated for grpRR,
i.e. each freqELB-th iteration (see also  \code{\link{fit_grpRR}})}

\item{calcELB}{whether to calculate the evidence lower bound (ELB) for grpRR (see also  \code{\link{fit_grpRR}})}

\item{th}{convergence threshold for the evidence lower bound (ELB) in grpRR (see also  \code{\link{fit_grpRR}})}

\item{n_rep}{number of reptitions with different random initilizations to be fit in grpRR (see also  \code{\link{fit_grpRR}})}

\item{verbose}{whether to print out intermediate messages during fitting}

\item{verbose_progress}{whether to print out details on the overall progress}

\item{compareGRridge}{whether to fit GRridge}

\item{include_nonfacQ}{whether to fit grpRR method with multivariate variational distributon be fitted (can be slow for large data sets)}

\item{compareSparseGroupLasso}{whether to fit sparse group Lasso}

\item{compareIPF}{whether to fit IPF-Lasso be fitted}

\item{compareGroupLasso}{whether to fit group Lasso}

\item{includeRF}{whether to fit random forest}

\item{compareAdaLasso}{whether to fit Lasso}

\item{include_varbvs}{whether to fit varbvs}

\item{include_nogamma}{whether to fit grpRR with same penalty factor but different sparity levels  per group}

\item{include_grpRR_SS_ungrouped}{whether to fit grpRR without group annotations}
}
\value{
List of fitted models and two data frames with
 coeffcients and penalty factors

a list with
 - a summaryList containing the complete individual fits
   as well as other statistics such as estimated coefficients,
    runtime, sparsity, intercept, penalty factors)
- the details on the data (sample size n, predictor number p,
 covariate annotation annot, group number G,
  feature and annotation names)
}
\description{
Function to fit a regression methods using serveral different methods.
}
\examples{
dat <- makeExampleData()
allFits <- runMethods(Xtrain=dat$X,
 ytrain=dat$y, annot=dat$annot)
}
