% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{runMethods}
\alias{runMethods}
\title{Run various regression methods}
\usage{
runMethods(Xtrain, ytrain, annot, family = "gaussian",
  intercept = TRUE, standardize = TRUE, beta0 = NULL,
  trueintercept = NULL, max_iter = 5000, freqELB = 10,
  calcELB = TRUE, th = 0.01, n_rep = 1, verbose = FALSE,
  verbose_progress = TRUE, includeGRridge = FALSE,
  include_grpRR_nonfacQ = FALSE, includeSparseGroupLasso = FALSE,
  includeIPF = FALSE, includeGroupLasso = FALSE, includeRF = FALSE,
  includeAdaLasso = FALSE, includeVarbvs = FALSE,
  include_grpRR_SS_nogamma = FALSE, include_grpRR_SS_ungrouped = FALSE)
}
\arguments{
\item{Xtrain}{design matrix with samples in rows and features in columns (n x p)}

\item{ytrain}{response vector of length n}

\item{annot}{factor of length p indicating group membership of each feature}

\item{family}{likelihood model for the response,
 either "gaussian" for linear regression
or "binomial" for logistic regression}

\item{intercept}{whether to include an intercept into the model}

\item{standardize}{whether to standardize the predictors to unit variance.
Note this does not affect GRridge and group Lasso  where standardization is default.}

\item{beta0}{true coefficients in the linear model if known,
NULL otherwise (default)}

\item{trueintercept}{true intercept in the linear model if known,
NULL otherwise (default)}

\item{max_iter}{maximum number of iterations for grpRR methods (see also  \code{\link{grpRR}})}

\item{freqELB}{frequency at which the evidence lower bound (ELB) is to be calculated for grpRR,
i.e. each freqELB-th iteration (see also  \code{\link{grpRR}})}

\item{calcELB}{whether to calculate the evidence lower bound (ELB) for grpRR (see also  \code{\link{grpRR}})}

\item{th}{convergence threshold for the evidence lower bound (ELB) in grpRR (see also  \code{\link{grpRR}})}

\item{n_rep}{number of reptitions with different random initilizations to be fit in grpRR (see also  \code{\link{grpRR}})}

\item{verbose}{whether to print out intermediate messages during fitting}

\item{verbose_progress}{whether to print out details on the overall progress}

\item{includeGRridge}{whether to fit GRridge}

\item{include_grpRR_nonfacQ}{whether to fit grpRR method with multivariate variational distributon be fitted (can be slow for large data sets)}

\item{includeSparseGroupLasso}{whether to fit sparse group Lasso}

\item{includeIPF}{whether to fit IPF-Lasso be fitted}

\item{includeGroupLasso}{whether to fit group Lasso}

\item{includeRF}{whether to fit random forest}

\item{includeAdaLasso}{whether to fit Lasso}

\item{includeVarbvs}{whether to fit varbvs}

\item{include_grpRR_SS_nogamma}{whether to fit grpRR with same penalty factor but different sparity levels  per group}

\item{include_grpRR_SS_ungrouped}{whether to fit grpRR without group annotations}
}
\value{
List of fitted models and two data frames with
 coeffcients and penalty factors

a list containg
\describe{
\item{summaryList}{list with the fitted models for each method that was included in the comparison, such as
\itemize{
\item grpRR_SS (sparse grpRR model with factorized variational distribution)
\item grpRR_FF (dense grpRR model with factorized variational distribution)
\item grpRR (dense grpRR model with nonfactorized variational distribution)
\item grpRR_SScutoff (as grpRR_SS where coefficients with posterior inclusion probabilities below 0.5 are set to zero)
\item grpRR_SS_nogamma (as grpRR_SS with a common slab precision across groups)
\item grpRR_SS_ungrouped (as grpRR_SS without group annotation)
\item Ridge (ridge regression)
\item Lasso (Lasso regression)
\item ElasticNet (elastic net regression)
\item varbvs (varbvs regression)
\item GroupLasso (group Lasso regression)
\item SparseGroupLasso (sparse group Lasso regression)
\item SparseGroupLasso (sparse group Lasso regression)
\item GRridge (GRridge regression)
\item NullModel (regression with the intercept term only)
\item TrueModel (regression with the true model coefficients (if known))
\item IPFLasso (IPF-Lasso regression)
\item adaptiveLasso (adaptive Lasso regression)
}
Each entry contains the
runtime, estimated penalty factors, coefficients, intercepts, sparsity level and the full output returned by the methods' call (out).
}
\item{groupnames}{groupnames from \code{annot}}
\item{varnames}{predictor names}
\item{family}{likelihoods model used}
\item{n}{number of samples}
\item{p}{number of features}
\item{G}{number of groups}
\item{annot}{annotation of features to groups as specified when calling \code{\link{runMethods}}}
}
}
\description{
Function to fit a linear or logistic regression model using serveral different methods.
}
\details{
This function fits a linear of logistic regression model to the data using various different methods.
grpRR is always included in its factorized form (both dense (grpRR_FF) and sparse (grpRR_SS)). In addition, a grpRR model is included with all
coefficients set to 0 whose posterior inclusion probaility (s) is below 50% (grpRR_SScutoff). As a comparison, ridge regression,
Lasso and elatic net as well as a intercept-only model are fitted. Other methods can be  included via the respective options, such as GRridge,
 non-facotized grpRR (grpRR), IPF-Lasso, sparse group Lasso, group Laso, adaptive Lasso, varbvs, random forest and
 grpRR without group annotations (grpRR_SS_ungrouped) or without different slab precisions (grpRR_nogamma).

 The fitted methods can be evaluated on test data using the function \code{\link{evaluateFits}}.
}
\examples{
dat <- makeExampleData()
allFits <- runMethods(Xtrain=dat$X, ytrain=dat$y, annot=dat$annot)
}
