% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{RunMethods}
\alias{RunMethods}
\title{RunMethods}
\usage{
RunMethods(Xtrain, ytrain, annot, beta0 = NULL, trueintercept = NULL,
  max_iter = 2000, intercept = T, plotit = F, standardize = T,
  verbose = F, compareGRridge = F, freqELB = 10, calcELB = T,
  include_nonfacQ = T, family = "gaussian", constantXcol = F,
  compareGroupLasso = T, includeRF = T, th = 1e-07, compareIPF = T,
  compareAdaLasso = T, n_rep = 10, include_varbvs = F)
}
\arguments{
\item{Xtrain}{Design matrix of size n x p}

\item{ytrain}{Response vector of size n}

\item{annot}{Factor of length p indicating group membership of each feature}

\item{beta0}{True coefficients in the linear model if known, NULL otherwise (default)}

\item{trueintercept}{True intercept in the linear model if known, NULL otherwise (default)}

\item{max_iter}{maximum number of iterations}

\item{intercept}{boolean, indicating wether to fit an intercept}

\item{plotit}{boolean, indicating wether to produce diagnositc plots}

\item{standardize}{boolean, indicating wether features for Ridge and Lasso fit should be standardized. Note this does not affect GRridge and grouplasso where standardization is default.}

\item{verbose}{boolean, indicating wether to print out intermediate messages during fitting}

\item{compareGRridge}{boolean, indicating wether to fit a GRridge model (might cause errors, look for stable version!)}

\item{freqELB}{determines frequency at which ELB is to be calculated, i.e. each feqELB-th iteration}

\item{calcELB}{boolean, indicating wether to calculate ELB}

\item{compareIPF}{boolean, indicating wether to fit a IPFLasso (might cause errors, look for stable version!)}

\item{compareAdaLasso}{boolean, indicating wether to fit an adpative lasso}
}
\value{
List of fitted models and two data frames with coeffcients and penalty factors
}
\description{
Function to run serveral different methods for high-dimensional regression
}
